+++ TrickyTowerDefense/Assets/Scripts/VO/IHitable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IHitable
 {
-    public void Hit(int damage);
+    public void Hit(float damage);
 }
